Downside information
The amount of information presented in the results have been reduced to about half of its original size. It now displays the bare minimum: badge number, employee name, date, amount of hours worked in that shift, and amount of hours missing in that shift. All entries in the results should have a value larger than 0 for the missing hours column. Entries that would've had 0 missing hours aren't included in the results.