Remove day filters for outer query
Since Sundays and Saturdays are already filtered out in the inner-most query, it's not necessary to use the same filters in the outer query. In fact, their inclusion in the outer query is detrimental to the execution time. With the filters in both places, the execution time was around 00:01:48. Now, the execution time is around a second.