Took out hard date+Add badge filter
The inner query had a line in the WHERE clause that looked for dates between January 1, 2021 and today. This was replaced with the user-created parameters. Some test badge numbers also appeared in the results, so the WHERE clause of the outer query now looks for badges that are larger than 0.